requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
alembic==1.11.1
    # via -r requirements.in
amqp==5.1.1
    # via kombu
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   httpcore
    #   starlette
    #   watchfiles
async-timeout==4.0.2
    # via redis
asyncpg==0.28.0
    # via -r requirements.in
bcrypt==4.0.1
    # via passlib
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via
    #   -r requirements.in
    #   flower
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
cffi==1.15.1
    # via cryptography
click==8.1.5
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cryptography==41.0.2
    # via pyjwt
dnspython==2.3.0
    # via email-validator
email-validator==2.0.0.post2
    # via
    #   fastapi
    #   fastapi-users
    #   pydantic
exceptiongroup==1.1.2
    # via
    #   anyio
    #   pytest
fastapi[all]==0.100.0
    # via
    #   -r requirements.in
    #   fastapi-cache2
    #   fastapi-users
fastapi-cache2[redis]==0.2.1
    # via -r requirements.in
fastapi-users[sqlalchemy]==12.1.0
    # via
    #   -r requirements.in
    #   fastapi-users-db-sqlalchemy
fastapi-users-db-sqlalchemy==6.0.0
    # via fastapi-users
flower==2.0.0
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.3
    # via httpx
httptools==0.6.0
    # via uvicorn
httpx==0.24.1
    # via fastapi
humanize==4.7.0
    # via flower
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.0.0
    # via pytest
itsdangerous==2.1.2
    # via fastapi
jinja2==3.1.2
    # via fastapi
kombu==5.3.1
    # via celery
makefun==1.15.1
    # via fastapi-users
mako==1.2.4
    # via alembic
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
orjson==3.9.2
    # via fastapi
packaging==23.1
    # via pytest
passlib[bcrypt]==1.7.4
    # via fastapi-users
pendulum==2.1.2
    # via fastapi-cache2
pluggy==1.2.0
    # via pytest
prometheus-client==0.17.1
    # via flower
prompt-toolkit==3.0.39
    # via click-repl
psycopg2-binary==2.9.6
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic[email]==2.0.2
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.1.2
    # via pydantic
pydantic-extra-types==2.0.0
    # via fastapi
pydantic-settings==2.0.1
    # via fastapi
pyjwt[crypto]==2.7.0
    # via fastapi-users
pytest==7.4.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.21.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   celery
    #   pendulum
python-dotenv==1.0.0
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.6
    # via
    #   fastapi
    #   fastapi-users
pytz==2023.3
    # via flower
pytzdata==2020.1
    # via pendulum
pyyaml==6.0
    # via
    #   fastapi
    #   uvicorn
redis==4.6.0
    # via fastapi-cache2
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy[asyncio]==2.0.18
    # via
    #   -r requirements.in
    #   alembic
    #   fastapi-users-db-sqlalchemy
starlette==0.27.0
    # via fastapi
tomli==2.0.1
    # via pytest
tornado==6.3.2
    # via flower
typing-extensions==4.7.1
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2023.3
    # via celery
ujson==5.8.0
    # via fastapi
uvicorn[standard]==0.22.0
    # via
    #   fastapi
    #   fastapi-cache2
uvloop==0.17.0
    # via uvicorn
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.19.0
    # via uvicorn
wcwidth==0.2.6
    # via prompt-toolkit
websockets==11.0.3
    # via uvicorn
